+++ tests/testthat/test-notes.R
@@ -7,6 +7,8 @@ library(yardstick)
 
 test_that("showing notes", {
   skip_if_not_installed("modeldata")
+  skip_if(rankdeficient_version)
+
   data(Chicago, package = "modeldata")
 
   base_wflow <-